Add files via upload
The manager is fully complete and has more or less full functionality. Now all that's left is to create another GUI for the customers. The plan is for them to be able to login with the newly created files, verify and then allow them to see their balance, deposit, withdraw and purchase things with the added automatic fee.